requirements.txt
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.8.30
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2
uuid==1.30
